Big changes for parallel-split and setting up communication groups
This commit makes big changes to the parallel-split setup by actually
splitting computation onto separate computational cores.  The new option
se_ftype=30 will accomplish this type of splitting.

This commit also introduces an experimental approach to handling issues
with writing output when the solution is solved on separate computational
cores.  This still needs to be tested given that the run will fail for certain
cases of core setups.

This commit also attempts to address an issue with setting up communication
groups in CAM or cases when dyn_npes < npes.  In the past there could be multiple
communication groups if the number of cores was significantly higher than the
chosen dyn_npes value.

[BFB] - Bit-For-Bit