Fixed code so that phys and dyn really are run in parallel.
Until now the code was not actually running physics and
dynamics in parallel, which was resulting in a performance
sink.  Now the code has been fixed so that dynamics is
executed on dynamics cores at the same time that
tphys_bc is executed on physics cores.

It is not known yet if this is BFB

See confluence for a more detailed description about these tags.