More work towards getting parallel-split phys/dyn coupling
This commit changed a few of the approaches to implementing the parallel
split algorithm for phys/dynamics coupling.  Namely, using the
applycamforcing subroutine rather than writing out all the steps in
stepon_run4.

This commit also introduces the dynamics state%q0 variable which
represents the dynamics state at the beginning of the physics step.  This
is necessary for calculting the tendencies from physics without
them being influenced by the changes to the dynamics state from
stepon_run3.

Lastly, this commit introduces the framework for solving physics and
dynamics on separate cores.  Although recent tests show that this is
still a broken implementation, so at this commit time it can be
considered that se_ftype=30 does not work.

Future commits will address this issue with running on separate cores.

[BFB] - Bit-For-Bit