First steps towards incorporating parallel-split phys/dyn coupling
This commit lays the groundwork for incorporating parallel-split
coupling of physics and dynamics.  It also introduces several metrics
for debugging current issues with this coupling approach.  Namely, it
has infrastructure for recording the model state following before and
after physics or dynamics have acted on the model (respectively).

[BFB] - Bit-For-Bit

See confluence for a more detailed description about these tags.