Fixed bug where qneg errors weren't actually recorded. Now are in ptend *********1*********2*********3*********4*********5*********6*********7**
This commit fixed an issue based on the assumption that the state variable
is constant throughout the physics calls.  This isn't true, in some cases
a dummy state (e.g. state1) is used.  The only true way to pass information
from a physics process to the actual state is through physics-tendencies,
(ptend).

[BFB] - Bit-For-Bit

See confluence for a more detailed description about these tags.