Consitent passing of tendencies from physics to dynamics
Added the option to pass the tracer tendencies (Qdp) instead of
the tracer state (Q) from physics to dynamics in p_d_coupling.

This is needed in order to run dynamics and physics concurrently.

It is controlled by setting the se_ftype to a negative value:
-2: same as ftype=2
-1: same as ftype=1
-10: same as ftype=0

[Non-BFB] - Non Bit-For-Bit

See confluence for a more detailed description about these tags.