collect hyperviscosity for the final step to occur after physics
*********1*********2*********3*********4*********5*********6*********7**
Moved the final horizontal hyperviscosity call to happen after the
application of physics tendencies for ftype=3.  Now the final application
of hyperviscosity occurs at the prim_run_subcycle layer, just before
the remap step.

This is needed to retain stability for the parallel-split version of
the model.

[BFB] - Bit-For-Bit   ! For normal phys/dyn coupling with limiter=4
[Non-BFB] - Non Bit-For-Bit ! for parallel-split

See confluence for a more detailed description about these tags.