Infrastructure for running physics and dynamics concurrently.
*********1*********2*********3*********4*********5*********6*********7**
This commit sets up the infrastructure for running most of dynamics
concurrently with most of physics.  If se_ftype=-3 then the dynamics
will be called directly following phys_run1, which holds the majority of
physics cost.  When dynamics is run on separate processors from physics
this will have the effect of running phys_run1 concurrently with
stepon_run3 (dynamics).

There is still a need to "finish" dynamics by applying the physics
tendencies, applying a final round of hyperviscosity, and then applying
vertical remapping.  That is done in cam_run3 using stepon_run3 with
mode=2.

[BFB] - Bit-For-Bit

See confluence for a more detailed description about these tags.