Complete transition to solving physics and dynamics on separate cores
Using the option phys_loadbalance = -2 cause the physics grid decompostion
to assign physics columns only to cores not already assigned dynamics
elements.  It is equivalent to phys_loadbalance = 2, the default option,
with this one distinction.

When using se_ftype=-3 this has the impact of running physics and dynamics
in parallel.  Any other ftype option results in a performance decrease.

[BFB] - Bit-For-Bit

See confluence for a more detailed description about these tags.