revise MPI commands in clubb_param_readnl
The original code declared a number of namelist variables
(e.g., clubb_C1) as real(kind=core_rknd) but used mpir8
in mpibcast commands.

In this revision, an integer variable "mpireal" is declared and used
by mpibcast. "mpireal" is assigned the value of mpir8 or mpir4 based
on the value of core_rknd.