Separate get_cloud_optics from set_cloud_optics
Separate the get_cloud_optics and set_cloud_optics routines, and call
both from radiation_tend. This makes it more obvious that we are
sub-sampling optics from the bands structure to gpoints, and allows
writing the more meaningful, intermediate cloud optics by bands from a
higher level rather than buried in the set_cloud_optics routine. This
will be important for MMF, where we may be looping over CRM columns and
do not want to send outputs to the history buffer each time through that
loop.