Modifies all tests to use PG2 grid
PG2 grid are faster than np4 grid as they have less columns. This
PR modifies tests in the mam4 testing suite as well as the auto
refactor test generation script. The baselines for the starting
point of the code are stored at:
/compyfs/e3sm_baselines/intel/mam4_org_v2_baselines/

We can use these baseline to compare against the refactored code.